Better plots for number of cores test
The throughput will be approximately constant if there's a lot of I/O
since then everything is waiting on that one I/O queue. But, if we have
more CPU bound processes, then the throughput will increase when there's
more processors to a point. Now we can see that. Amdahl's law.