Added SPN queue algorithm
This required that I provide some method of accessing the PCB directly
from within the queue, so it can look at the remaining execution times.
I did this by just making the queues store references to the PCBs
directly rather than the pid which then is looked up in the processTable
dictionary.