Discrete event simulation
Don't just increment the clock by 1 but jump to the next event that will
occur. This really complicates things, but it makes them much faster (if
debugTiming = True, it went from 37 s real down to 13 s). However, I
wanted to make sure that everything was exactly the same, so did lots of
comparisons with the old code.

Also added HRRN, but wanted to fix the discrete event simulation before
testing that.